SURA: EmployeeExpense  Model and Factory

+++ application/src/main/java/com/sura/expenses/employee/command/CommandEmployee.java
@@ -12,5 +12,5 @@ public class CommandEmployee {
     private Long employeeId;
     private String employeeName;
     private String expenseDate;
-    private Double totalExpense;
+    private Integer totalExpense;
 }
